Day 13
I wasted quite a bit of time in Part 1 because I didn't realise that you only needed to print the number of dots after the FIRST fold, not ALL of the folds.
At least it helped me get through Part 2 quicker, as i didn't need to add code for that.